pyproject.toml
[project]
name = "shape-to-solid"
version = "0.1.0"
description = "A project to convert shapes to solids."
authors = [{ name = "Your Name", email = "you@example.com" }]
dependencies = [
    "crewai>=1.9.3",
    "mcp-server-git>=2026.1.14",
    "python-dotenv>=1.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/shape_to_solid"]


